Refactor DataCollectionScreen by extracting components and previews to separate files
