+++ app/src/main/java/org/groundplatform/android/ui/datacollection/DataCollectionScreen.kt
@@ -16,7 +16,6 @@
 package org.groundplatform.android.ui.datacollection
 
 import androidx.compose.animation.core.animateFloatAsState
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
@@ -39,24 +38,15 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.layout.onGloballyPositioned
 import androidx.compose.ui.layout.positionInWindow
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.IntOffset
-import androidx.compose.ui.viewinterop.AndroidView
-import androidx.core.view.doOnLayout
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import androidx.viewpager2.widget.ViewPager2
-import kotlinx.serialization.json.JsonObject
 import org.groundplatform.android.R
 import org.groundplatform.android.ui.components.ConfirmationDialog
 import org.groundplatform.android.ui.components.Toolbar
-import org.groundplatform.domain.model.job.Job
-import org.groundplatform.domain.model.locationofinterest.LoiReport
-import org.groundplatform.ui.theme.AppTheme
 
 /**
  * The main screen for data collection, coordinating the task sequence and host UI.
@@ -104,37 +94,6 @@ fun DataCollectionScreen(
 }
 
 /**
- * Hosts the ViewPager2 that renders individual task screens.
- *
- * @param uiState The current UI state.
- * @param fragment The fragment hosting this screen.
- */
-@Composable
-private fun DataCollectionViewPager(
-  uiState: DataCollectionUiState,
-  fragment: DataCollectionFragment,
-) {
-  AndroidView(
-    factory = { context ->
-      ViewPager2(context).apply {
-        isUserInputEnabled = false
-        offscreenPageLimit = 1
-      }
-    },
-    update = { viewPager ->
-      if (uiState is DataCollectionUiState.Ready) {
-        val currentAdapter = viewPager.adapter as? DataCollectionViewPagerAdapter
-        if (currentAdapter == null || currentAdapter.tasks != uiState.tasks) {
-          viewPager.adapter = fragment.viewPagerAdapterFactory.create(fragment, uiState.tasks)
-        }
-        viewPager.doOnLayout { viewPager.setCurrentItem(uiState.position.absoluteIndex, false) }
-      }
-    },
-    modifier = Modifier.fillMaxSize(),
-  )
-}
-
-/**
  * The layout content of the data collection screen, including toolbar, pager content, and progress
  * bar.
  *
@@ -162,22 +121,17 @@ fun DataCollectionContent(
       ) {
         when (uiState) {
           is DataCollectionUiState.Loading -> {
-            CircularProgressIndicator(
-              modifier = Modifier.align(Alignment.Center).testTag("loading_indicator")
-            )
+            LoadingContent(Modifier.align(Alignment.Center))
           }
           is DataCollectionUiState.Error -> {
-            Text(
-              text = "Error: ${uiState.code}",
-              modifier = Modifier.align(Alignment.Center).testTag("error_message"),
-              color = MaterialTheme.colorScheme.error,
-            )
+            ErrorContent(uiState.code, Modifier.align(Alignment.Center))
           }
           is DataCollectionUiState.Ready -> {
-            pagerContent()
-            DataCollectionProgressBar(
-              uiState.position,
-              progressPositionY = footerVerticalPosition - boxPositionY,
+            ReadyContent(
+              position = uiState.position,
+              footerVerticalPosition = footerVerticalPosition,
+              boxPositionY = boxPositionY,
+              pagerContent = pagerContent,
             )
           }
           is DataCollectionUiState.TaskSubmitted -> {
@@ -187,6 +141,31 @@ fun DataCollectionContent(
       }
     }
   }
+}
+
+@Composable
+private fun LoadingContent(modifier: Modifier = Modifier) {
+  CircularProgressIndicator(modifier = modifier.testTag("loading_indicator"))
+}
+
+@Composable
+private fun ErrorContent(code: DataCollectionErrorCode, modifier: Modifier = Modifier) {
+  Text(
+    text = "Error: $code",
+    modifier = modifier.testTag("error_message"),
+    color = MaterialTheme.colorScheme.error,
+  )
+}
+
+@Composable
+private fun ReadyContent(
+  position: TaskPosition,
+  footerVerticalPosition: Float,
+  boxPositionY: Float,
+  pagerContent: @Composable () -> Unit,
+) {
+  pagerContent()
+  DataCollectionProgressBar(position, progressPositionY = footerVerticalPosition - boxPositionY)
 }
 
 /**
@@ -239,84 +218,4 @@ private fun DataCollectionProgressBar(position: TaskPosition, progressPositionY:
         .offset { IntOffset(0, progressPositionY.toInt()) }
         .testTag("progress_bar"),
   )
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun DataCollectionContentLoadingPreview() {
-  AppTheme {
-    DataCollectionContent(
-      uiState = DataCollectionUiState.Loading,
-      footerVerticalPosition = 100f,
-      onCloseClicked = {},
-    ) {
-      Box(modifier = Modifier.fillMaxSize().background(Color.LightGray)) {
-        Text(text = "Pager Content Area", modifier = Modifier.align(Alignment.Center))
-      }
-    }
-  }
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun DataCollectionContentErrorPreview() {
-  AppTheme {
-    DataCollectionContent(
-      uiState =
-        DataCollectionUiState.Error(
-          code = DataCollectionErrorCode.NO_VALID_TASKS,
-          cause = Error("Some error"),
-        ),
-      footerVerticalPosition = 100f,
-      onCloseClicked = {},
-    ) {
-      Box(modifier = Modifier.fillMaxSize().background(Color.LightGray)) {
-        Text(text = "Pager Content Area", modifier = Modifier.align(Alignment.Center))
-      }
-    }
-  }
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun DataCollectionContentPreview() {
-  AppTheme {
-    DataCollectionContent(
-      uiState =
-        DataCollectionUiState.Ready(
-          surveyId = "survey1",
-          job = Job(id = "job1", name = "Test Job"),
-          loiName = "Test LOI",
-          tasks = emptyList(),
-          isAddLoiFlow = false,
-          currentTaskId = "task1",
-          position = TaskPosition(0, 1, 2),
-        ),
-      footerVerticalPosition = 2000f,
-      onCloseClicked = {},
-    ) {
-      Box(modifier = Modifier.fillMaxSize().background(Color.LightGray)) {
-        Text(text = "Pager Content Area", modifier = Modifier.align(Alignment.Center))
-      }
-    }
-  }
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun DataCollectionContentCompletePreview() {
-  AppTheme {
-    DataCollectionContent(
-      uiState =
-        DataCollectionUiState.TaskSubmitted(
-          loiReport = LoiReport(loiName = "Point A", geoJson = JsonObject(mapOf()))
-        ),
-      footerVerticalPosition = 2000f,
-      onCloseClicked = {},
-    ) {
-      Box(modifier = Modifier.fillMaxSize().background(Color.LightGray)) {
-        Text(text = "Pager Content Area", modifier = Modifier.align(Alignment.Center))
-      }
-    }
-  }
 }
